[Player] Move business logic to new PlayerContainer
