+++ src/components/App.jsx
@@ -1,13 +1,13 @@
 import React from 'react';
 
-import { Header, TeaserList, AudioSettings, Player } from 'app/components';
+import { Header, TeaserList, AudioSettings, PlayerContainer } from 'app/components';
 
 const App = () => (
     <div>
         <Header />
         <AudioSettings />
         <TeaserList />
-        <Player />
+        <PlayerContainer />
     </div>
 );
